Publishing field and communication channel fields now are working properly
