+++ apps/legal-gazette-application-web/src/lib/forms/common-form.ts
@@ -32,6 +32,7 @@ export const commonForm = ({
       signatureLocation: fields.signatureLocation,
       signatureDate: fields.signatureDate,
       publishingDates: fields.publishingDates,
+      communicationChannels: fields.communicationChannels,
     },
   },
 })
